chore: rename CpsAssets folder to CpsAsset
